v4l2n: do not store images into memory unless -o is given
If -o (output filename) was not given, v4l2n saved images into
memory even if not writing them to disk. Change the behaviour
to discard immediately any captured images unless -o option is
given. The option has to be given in the beginning of command
line to save all captured frames from the beginning.