v4l2n: clear exception handling (longjump) after fork()
When executing another program after a fork, if the execution failed,
error() was called which jumped to exception handling which in turn
saved all buffers to files.

If the program execution fails, we do not want to save buffers or
anything from the forked process of v4l2n. Therefore reset the longjump
to do nothing.

Signed-off-by: Tuukka Toivonen <tuukka.toivonen@intel.com>